Update roadmap.md
Added sections to William Areas in Hardware, including all needed materials to make the rolling blinds motor. I also made changes in software needed like Arduino Uno's to import information onto. I made changes in errors in the code and reasoning for them, which was because there was no software to import stored data onto, and the way to fix it would be to buy an Arduino mini so it can be properly transfered.